+++ Parley/Parley/Views/QuestBrowserWindow.axaml.cs
@@ -5,6 +5,7 @@ using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
 using Avalonia.Controls;
+using Avalonia.Input;
 using Avalonia.Interactivity;
 using DialogEditor.Models;
 using DialogEditor.Services;
@@ -512,5 +513,24 @@ namespace DialogEditor.Views
         {
             Close(null);
         }
+
+        #region Title Bar Events
+
+        private void OnTitleBarPointerPressed(object? sender, PointerPressedEventArgs e)
+        {
+            if (e.GetCurrentPoint(this).Properties.IsLeftButtonPressed)
+            {
+                BeginMoveDrag(e);
+            }
+        }
+
+        private void OnTitleBarDoubleTapped(object? sender, RoutedEventArgs e)
+        {
+            WindowState = WindowState == WindowState.Maximized
+                ? WindowState.Normal
+                : WindowState.Maximized;
+        }
+
+        #endregion
     }
 }